refactor: restructure header components and improve navigation layout
