+++ src/components/common/HeadingText.tsx
@@ -3,7 +3,7 @@ import styles from "../../styles/abstracts/typography.module.scss";
 
 type Props = {
   className?: string;
-  as?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "p" | "a";
+  as?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "p" | "a" | "span";
   size?: keyof typeof typographySizes;
   title?: string;
   children?: any;
