+++ src/components/header/Header.tsx
@@ -1,0 +1,45 @@
+import { useState } from "react";
+import styles from "../../styles/components/common/Header.module.scss";
+import { CartOverlay } from "../cart/CartOverlay.tsx";
+import { useNavigate } from "react-router-dom";
+import clsx from "clsx";
+import { Logo, NavLinks, IconLinks } from "./index";
+
+type Props = {
+  className?: string;
+};
+
+const Header = ({ className }: Props) => {
+  const [isCartOpen, setIsCartOpen] = useState(false);
+  const navigate = useNavigate();
+
+  const handleCartClick = (e: React.MouseEvent) => {
+    e.preventDefault();
+    setIsCartOpen(true);
+  };
+
+  const handleProfileClick = (e: React.MouseEvent) => {
+    e.preventDefault();
+    void navigate("/login");
+  };
+
+  return (
+    <>
+      <nav
+        className={clsx(styles.header, className)}
+        aria-label="Main Navigation"
+      >
+        <Logo />
+        <NavLinks />
+        <IconLinks
+          onCartClick={handleCartClick}
+          onProfileClick={handleProfileClick}
+        />
+      </nav>
+
+      <CartOverlay isOpen={isCartOpen} onClose={() => setIsCartOpen(false)} />
+    </>
+  );
+};
+
+export default Header;
