+++ src/components/header/index.ts
@@ -1,0 +1,4 @@
+export { default as IconLinks } from "./IconLinks";
+export { default as NavLinks } from "./NavLinks";
+export { default as Logo } from "./Logo";
+export { default as Header } from "./Header";
